Add files via upload
I developed a dynamic gym website using HTML and CSS, incorporating advanced techniques such as positioning and utilizing classes and IDs. The website features a visually appealing design with a quick demo, contact us, and about us pages. It includes a distinctive logo and a convenient form for users to book a demo session effortlessly.